extracted eval func of point segmentation into a separate file
